+++ administrator/components/com_content/controllers/articles.php
@@ -105,63 +105,6 @@ class ContentControllerArticles extends JControllerAdmin
 	}
 
 	/**
-	 * Method to discard shared drafts.
-	 *
-	 * @return  void
-	 *
-	 * @since   __DEPLOY_VERSION__
-	 */
-	public function discardDraft()
-	{
-		// Check for request forgeries
-		JSession::checkToken() or jexit(JText::_('JINVALID_TOKEN'));
-
-		$ids = $this->input->get('cid', array(), 'array');
-
-		// Access checks.
-		foreach ($ids as $i => $id)
-		{
-			if (!JFactory::getUser()->authorise('core.edit.state', 'com_content.article.' . (int) $id))
-			{
-				// Prune items that you can't change.
-				unset($ids[$i]);
-
-				JError::raiseNotice(403, JText::_('JLIB_APPLICATION_ERROR_EDITSTATE_NOT_PERMITTED'));
-			}
-		}
-
-		if (empty($ids))
-		{
-			JError::raiseWarning(500, JText::_('JERROR_NO_ITEMS_SELECTED'));
-		}
-		else
-		{
-			// Get the model.
-			$model = $this->getModel('articles');
-
-			// Publish the items.
-			if (!$model->discardDraft($ids))
-			{
-				JError::raiseWarning(500, $model->getError());
-			}
-		}
-
-		// Publish the items.
-		if (!$model->discardDraft($ids))
-		{
-			JError::raiseWarning(500, $model->getError());
-
-			return false;
-		}
-
-		$message = JText::plural('COM_CONTENT_N_ITEMS_UNSHARED', count($ids));
-		$this->setRedirect(JRoute::_('index.php?option=com_content&view=shared', false), $message);
-
-	}
-
-
-
-	/**
 	 * Proxy for getModel.
 	 *
 	 * @param   string  $name    The model name. Optional.
