+++ administrator/components/com_content/models/articles.php
@@ -9,8 +9,6 @@
 
 defined('_JEXEC') or die;
 
-use Joomla\Utilities\ArrayHelper;
-
 /**
  * Methods supporting a list of article records.
  *
@@ -391,46 +389,5 @@ class ContentModelArticles extends JModelList
 		}
 
 		return $items;
-	}
-
-	/**
-	 * Method to discard Shared Drafts,
-	 * 
-	 * @param   array  $pks  A prefix for the store id.
-	 *
-	 * @return  mixed  An array of data items on success, false on failure.
-	 *
-	 * @since   __DEPLOY_VERSION__
-	 */
-	public function discardDraft($pks)
-	{
-		// Sanitize the ids.
-		$pks = ArrayHelper::toInteger($pks);
-
-		if (empty($pks))
-		{
-			$this->setError(JText::_('COM_CONTENT_NO_ITEM_SELECTED'));
-
-			return false;
-		}
-
-		try
-		{
-			$db = $this->getDbo();
-			$query = $db->getQuery(true)
-				->delete($db->quoteName('#__share_draft'))
-				->where($db->quoteName('articleId') . ' IN (' . implode(',', $pks) . ')');
-			$db->setQuery($query);
-			$db->execute();
-		}
-		catch (Exception $e)
-		{
-			$this->setError($e->getMessage());
-
-			return false;
-		}
-
-		return true;
-
 	}
 }
